+++ crates/parser-common/CHANGELOG.md
@@ -1,5 +1,11 @@
+# socketioxide-parser-common 0.17.2
+* fix(security): implement `ParseConfig` rules to fix unlimited binary data attachments which could lead
+to denial of service. See [Github Advisory](https://github.com/Totodore/socketioxide/security/advisories/GHSA-55mf-67qm-4wpg).
+* deps: bump `socketioxide-core` to 0.19
+
 # socketioxide-parser-common 0.17.1
 * deps: bump `socketioxide-core` to 0.18
+
 # socketioxide-parser-common 0.17
 * deps: bump `socketioxide-core` to 0.17
 * MSRV: rust-version is now 1.86 with edition 2024