+++ crates/socketioxide-core/CHANGELOG.md
@@ -1,3 +1,6 @@
+# socketioxide-core 0.19.0
+* feat(*breaking*): Add a new `ParseConfig` arg to parser `decode_{str/bin}` methods. 
+
 # socketioxide-core 0.18.2
 * feat: add new `Volatile` broadcast flag